removed map private check, uint cast to pa in proc.c

+++ xv6-public/proc.c
@@ -307,7 +307,7 @@ fork(void)
         panic("kalloc");
       }
 
-      if(mappages(np->pgdir, (void *) curr_vma->start, PGSIZE, pa, curr_vma->prot | PTE_U)!=0){
+      if(mappages(np->pgdir, (void *) curr_vma->start, PGSIZE, (uint)pa, curr_vma->prot | PTE_U)!=0){
         kfree(pa);
         np->killed = 1;
       };
